refactor: move history_note to history package
