+++ pkg/storage/internal/history/history_note.go
@@ -1,4 +1,4 @@
-package entity
+package history
 
 var InternalizeActionHistoryNote = "internalizeAction"
 
@@ -7,3 +7,4 @@ func UserIDHistoryAttr(userID int) map[string]any {
 		"userId": userID,
 	}
 }
+
